Updating models

+++ src/main/java/com/pismo/test/model/OperationsTypes.java
@@ -1,8 +1,6 @@
 package com.pismo.test.model;
 
 import lombok.Data;
-import lombok.Getter;
-import lombok.Setter;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -15,6 +13,6 @@ public class OperationsTypes {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
+    private Long id;
     private Description description;
 }
